requirements.txt
Streamlit
pandas>=1.1.3
numpy>=1.19.2
matplotlib>=3.2.2
seaborn==0.11.0
plotly>=4.4
requests==2.25.1
Pillow==8.1.0
scikit_learn>=0.24.1
lightgbm==3.2.1
shap>=0.39
lime==0.2.0.1
streamlit==0.52.2
toml==0.10.0